decision_algoritm/sma/requirements.txt
python-binance
python-dotenv
mysql-connector-python
requests
pytest
coverage
